Split lexer syntax example tests
